refactor(localhost): change subject structure
overall structure modified to have more modulary and clarity.
this is still a work in progress and far to be complete.
at the moment few parts were removed/replaced, focus being in reorganizing what already exists.
future commits should focus on adding restrictions/information for clarity and to better define the work to be done by the student.