Update README.md
Create function named `ionOut`, that receives a `string` and returns an array. Array should only include words containing `'ion'` followed by `'t'`. Words should be returned without the `'ion'` part. 

It is very confusing and misleading, it is a must to make sure task has correct discription to make sure task is done properly at first try, no need to find out that you have to have an empty array when 'tion' part is not in.